% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExecutor.R
\name{getExecutor}
\alias{getExecutor}
\title{Get the parallel executor}
\usage{
getExecutor()
}
\value{
An external pointer to be passed to \code{Rtatami::set_executor}.
}
\description{
Get the executor object for safe execution of R code in parallel sections.
}
\examples{
getExecutor()

}
\author{
Aaron Lun
}
