% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichNonZero.R
\name{whichNonZero}
\alias{whichNonZero}
\title{Find non-zero entries of a matrix}
\usage{
whichNonZero(x, ...)
}
\arguments{
\item{x}{A numeric matrix-like object, usually sparse in content if not in representation.}

\item{...}{Further arguments, ignored.}
}
\value{
A list containing \code{i}, an integer vector of the row indices of all non-zero entries;
\code{j}, an integer vector of the column indices of all non-zero entries;
and \code{x}, a (usually atomic) vector of the values of the non-zero entries.
}
\description{
This function is soft-deprecated; users are advised to use \code{\link[SparseArray]{nzwhich}} and \code{\link[SparseArray]{nzvals}} instead.
}
\examples{
x <- Matrix::rsparsematrix(1e6, 1e6, 0.000001)
out <- whichNonZero(x)
str(out)

}
\seealso{
\code{\link{which}}, obviously.
}
\author{
Aaron Lun
}
