% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_beachmat.R
\name{use_beachmat}
\alias{use_beachmat}
\title{Use beachmat}
\usage{
use_beachmat()
}
\description{
Creates \code{src/} and adds needed packages to \code{DESCRIPTION}.
}
\details{
For more details see the 'Linking' vignette, available by running
\code{vignette("linking", package = "beachmat")}.
}
